requirements.txt
openai
discord.py
python-decouple
deep-translator